livemerge: Fix handling of last extend failure
If we start 10 extends because of a timeout, and then one of the older
extends fail, we would fail the entire merge, instead of waiting until
the last extend attempt complete or times out.

Now we pass the attempt number to the callback, so we can consider the
real attempt number, and fail the entire merge only if the last attempt
has failed.

Change-Id: I4af66497c16bd765c1c7bb6905d286956ef4f289
Signed-off-by: Nir Soffer <nsoffer@redhat.com>