backup.py: remove parent_checkpoint_id validation
Libvirt support the options to redefine single checkpoint and not the
entire checkpoints chain, this can be done because now all the
checkpoints in the chain are active and not just the leaf checkpoint.

Supporting single checkpoint redefinition means that there is no use
for the parent checkpoint ID anymore and the validation for it is not
needed.

This patch removes the validation for the parent_checkpoint_id.

Change-Id: I669d781b18a8fdc1cdf8c95abefe29a6a3a7fcf1
Bug-Url: https://bugzilla.redhat.com/1950752
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>