sp: Improve canceling upgrade during stopSpm
When stopping the SPM, we try to cancel upgrade domain. If the upgrade
already started, we wait until the upgrade is completed. Rename the
function to reflect the actual intent, add a docstring, and improve the
log message. The new log is using INFO level so we have more visibility
in this critical flow.

Change-Id: Id06493a3420883767ee229585a83a22c8bc70666
Related-to: https://bugzilla.redhat.com/1952345
Signed-off-by: Nir Soffer <nsoffer@redhat.com>