sp: Rename UNSECURE to INSECURE
And update the methods managing secure mode to modern style. This is not
consistent with other code in sp.py, but this inconsistency has a
purpose - it helps to tell which code is trustable and which is not.

Here are example logs when host become the SPM:

2021-04-27 23:40:02,707+0300 INFO  (tasks/0) [storage.StoragePool]
Switching storage pool fb7797bd-3cca-4baa-bcb3-35e8851015f7 to SECURE
mode (sp:130)

And when host stops being the SPM:

2021-04-27 23:40:14,195+0300 INFO  (jsonrpc/3) [storage.StoragePool]
Switching storage pool fb7797bd-3cca-4baa-bcb3-35e8851015f7 to INSECURE
mode (sp:137)

Change-Id: I64014ade5af4289fc17ad6d0b5d7e04469386b13
Related-to: https://bugzilla.redhat.com/1952345
Signed-off-by: Nir Soffer <nsoffer@redhat.com>