virt: refresh disk on destination during migration
VM's disks can change on underlying storage while migration is running.
Such example is e.g. extending the disk. The disk is extended on the
source host, but the change is not visible on the destination host.

Refresh the disk being extended first on the destination host during
migration, to be sure that the disk won't become corrupted if the
migration process is finihed and VM on the destination is resumed by
libvirt.

Change-Id: I9f3739c7c64837b3730149216166de4a65dc4c48
Bug-Url: https://bugzilla.redhat.com/1883399
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>