qga: do not poll libvirt for all guest info
If our logic decides that there is nothing to query qemu-ga for
(types == 0), which can happen several times every minute, we have to
bail out and not pass this to libvirt API. Calling libvirt API with
types = 0 will query all available information. This is not something we
want as it burdens the guest unnecessarily.

Bug-Url: https://bugzilla.redhat.com/1944495
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Change-Id: Ia18e7ec5e5de22a2980066d6a80bb72fcfef65de