tests: add migration thread into live merge test VM
Add migration thread into live merge RunningVM. This is needed by stuff
anyhow touching migration. E.g. extending disk, which is used in tests
in this module, uses VM.after_volume_extension() as a callback. In
followup patch callback will include a check if migration is ongoing,
which would cause of the tests to fail.

Change-Id: I6bf99fa23a437bb2097d673e3d4568e8568ca2dc
Bug-Url: https://bugzilla.redhat.com/1883399
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>