clusterlock: Panic if we lost leases silently
In sanlock.acquire() or sanlock.release(), sanlock client is
communicating with sanlock daemon using the process fd (slkfd=) instead
of using new connection (pid=). If sanlock daemon closed the socket, the
call will fail with SanlockExcpetion(EPIPE).

If we are holding some leases when this error happens, this is a
catastrophic event meaning that sanlock released our leases behind our
back. This could lead to split brain and data corruption, similar to
bug[1].

In SANLock.aquire() we recovered from this error *silently* by
registering a new socket with sanlock daemon and retrying the call. This
hides the catastrophic event, making it impossible to debug. In
SANLock.release() we were raising the error, which is little better,
leaving some evidence in the log.

When sanlock releases our leases behind our back, there is no good way
to recover, so the best thing we can do is panic. Vdsm will lose the SPM
role, and child processes started by vdsm will be killed.

Add SANLock._lease_count counter, increased when acquiring a lease, and
decreased when releasing a lease. When we detect EPIPE on the sanlock
process fd, and the lease count is non-zero, we panic with this message:

    Sanlock process fd was closed while holding 3 leases: ...

[1] https://bugzilla.redhat.com/1952345

Change-Id: I388ff66627de341edca42bb2614da68655d6da47
Signed-off-by: Nir Soffer <nsoffer@redhat.com>