net: Remove dracut conf file
The dracut conf file was omitting the ifcfg and
clevis modules because it was causing issues with
bridged networking. As consequence users that wanted
to run auto-decrypt via clevis were failing to do so.

The original issue is resolved by the ability
of NetworkManager to run in dracut that is enabled
by default since EL8.3 [0].

[0] https://bugzilla.redhat.com/1627820

Change-Id: Iff3b85c24c2ec6736aea8cd9c35e8cb366b50420
Bug-Url: https://bugzilla.redhat.com/1955571
Bug-Url: https://bugzilla.redhat.com/1959945
Signed-off-by: Ales Musil <amusil@redhat.com>