clusterlock: Rename inquire() to inspect()
We use the name inquire() for querying a lease that may be held by
another host. This is a heavy operation using 3 sanlock APIs accessing
storage. Sanlock has an inquire() API, which is a lightweight operation
accessing sanlock daemon state on the current host, and returning list
of leases owned by the caller program of by another program.

Rename our special version to inspect(), to make room for
sanlock.inquire().

Change-Id: Ic3a137475ac5c874b6df9dceca22cfd77b961dee
Bug-Url: https://bugzilla.redhat.com/1961752
Signed-off-by: Nir Soffer <nsoffer@redhat.com>