sp: Don't hide error in stopSpm
When unrecoverable error occur during stopSpm we mark the operation as
failed, and continue with the next call. Finally we panic with a generic
message:

    2021-05-20 16:08:38,961+0300 ERROR (jsonrpc/2) [root] Panic:
    Unrecoverable errors during SPM stop process. (panic:29)

But since we did not log the error, we don't have a clue what was the
unrecoverable error.

Now we log an exception for every error during stopSpm.

Change-Id: I2362576d9443cb6d4145e8fe0399ac7881f18a80
Signed-off-by: Nir Soffer <nsoffer@redhat.com>