sp: Failure to stop mail monitor is critical
Previously we did not treat failure to stop the mail monitor as fatal
error. In this case we could have running mail monitor on a non-spm
host. This is likely to corrupt lvm metadata.

Treat this error as critical error. If mail monitor could not be
stopped, vdsm will panic.

Change-Id: I9ceaa567b10804c0eb8c0ce078beb50a27533f5a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>