sp: Wait until SPM mail monitor is stopped
When stopping the SPM, we did not wait until the mail monitor was
stopped before releasing the cluster lock. Stopping the mailbox sets a
flag and return. Some mailbox worker threads may be busy handling
extension request. If a new SPM is started, we can have 2 hosts trying
to serve extension requests, which is likely to corrupt lvm metadata.

Now we wait up to 60 seconds for mail monitor termination. If the
operation times out we treat it as fatal failure and panic.

We expect to see this log in stopSpm flow:

    2021-05-20 17:10:24,173+0300 INFO  (mailbox-spm) [storage.MailBox.SpmMailMonitor]
    SPM_MailMonitor - Incoming mail monitoring thr ead stopped (mailbox:765)

If stopping the mail monitor times out, we expect to see:

    2021-05-20 19:28:54,562+0300 ERROR (jsonrpc/4) [storage.StoragePool] Timeout stopping SPM mail
    monitor (sp:456)

And vdsm will panic:

    2021-05-20 19:28:54,568+0300 ERROR (jsonrpc/4) [root] Panic: Unrecoverable errors during SPM
    stop process. (panic:31)

Change-Id: I7e7d4c8f68b04bd7539a31f6d1a6dd05ef8ddb06
Signed-off-by: Nir Soffer <nsoffer@redhat.com>