panic: Shutdown logging before killing process group
When handling unrecoverable error we typically log some error or
exception, and then the panic message. For example in this case we panic
while releasing the SPM lease:

    2021-05-20 17:10:24,173+0300 INFO  (jsonrpc/7) [storage.SANLock] Releasing Lease(name='SDM',
    path='/dev/8ef0cc64-77c5-49e0-a0f5-03cee9bbee20/leases', offset=1048576) (clusterlock:602)

    2021-05-20 17:10:29,957+0300 INFO  (MainThread) [vds] (PID: 853041) I am the actual vdsm
    4.40.70.10.git91384f25a host4 (4.18.0-3
    05.el8.x86_64) (vdsmd:158)

But the logger thread was killed before writing the error and panic
message to the log.

Shutdown logging before killing the process group. This close all
handlers, ensuring that pending messages are logged.

Change-Id: I96f2fd9890414437188ae3b3ba8bf29fc79bb60c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>