spwd: Introduce storage pool watchdog
Add watchdog watching the master storage domain cluster lock after the
SPM lease was acquired, and before it was released. The watchdog will
panic on issues with the SPM lease, causing the host to lose the SPM
role, and killing all child processes. This will trigger selection of
new SPM.

The watchdog panics in these cases:
- The SPM lease is not reported by sanlock.
- The SPM lease has unexpected disk (path, offset)
- Inquiring sanlock failed with temporary error more than 3 times
- Inquiring sanlock failed with unexpected error

We should also verify the lease version, but turns out that engine does
not pass the SPM lease version, so we don't have it when starting the
SPM.

The watchdog should be used only when the master domain cluster lock
reports the new "supports_inquire" property.

New logs introduced by this change:

Starting the watchdog:

    17:24:35,677 INFO    (MainThread) [storage.spwd] Start watching cluster
    lock Lease(lockspace='master-domain-uuid', resource='SDM',
    disk=('/master/leases', 1048576), version=1) (spwd:79)

Stopping the watchdog:

    17:24:36,179 INFO    (MainThread) [storage.spwd] Stop watching cluster
    lock Lease(lockspace='master-domain-uuid', resource='SDM',
    disk=('/master/leases', 1048576), version=1) (spwd:88)

In debug level, we log each successful check:

    17:24:35,879 DEBUG   (spwd) [storage.spwd] Found cluster lock
    {'lockspace': 'master-domain-uuid', 'resource': 'SDM', 'version': 1,
    'disks': [('/master/leases', 1048576)]} (spwd:149)

If the watchgod fails with a temporary error, we log a warning:

    17:35:54,378 WARNING (spwd) [storage.spwd] Error (2/3) checking cluster
    lock Lease(lockspace='master-domain-uuid', resource='SDM',
    disk=('/master/leases', 1048576), version=1) (spwd:119)

If the watchdog panics, we will see a panic exception message like:

    Panic: Invalid cluster lock disk expected=Lease(
    lockspace='master-domain-uuid', resource='SDM',
    disk=('/master/leases', 1048576))
    actual={'lockspace': 'master-domain-uuid', 'resource': 'SDM',
    'version': 2, 'disks': [('/master/leases', 10485760)]}

Change-Id: Ied3efef9af84dd0d05ec54801516b799a0839043
Bug-Url: https://bugzilla.redhat.com/1961752
Signed-off-by: Nir Soffer <nsoffer@redhat.com>