volumemetadata: rename attribute to increase consistency
Parent attribute is not consistent, there is no reason to name it
'puuid' since we present the value as 'parent' instead, e.g. in
dump output.

The only place 'puuid' is used is storage format key (sc.PUUID const.)

Renaming it for consistency to allow cleaner code in followup patches.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1870887
Change-Id: I026fa4d3696c737cea7e3e4fed3a55989f543c7c