fakesanlock: Simulate EBUSY
Sanlock may fail resource operations with EBUSY if a client command is
busy performing operation in the thread pool. This should not happen in
vdsm since we serialize calls to acquire(), release(), and inquire(),
but I'm not 100% sure this is not possible.

Add "busy" flag to resource, so we can simulate a failure using EBUSY
error.

Change-Id: I88ef16c91ae8d30592a72cbbad9af13aeaba938e
Bug-Url: https://bugzilla.redhat.com/1961752
Signed-off-by: Nir Soffer <nsoffer@redhat.com>