virt: Switch to POWERING_DOWN status only from UP
Logic at many places is dependent on Vm.lastStatus.  POWERING_DOWN VM
status is mostly informative and used mainly in interaction with a
guest agent.

When asking a VM to shut down, it's dangerous to change its status to
POWERING_DOWN status unconditionally.  For instance, consider the
following scenario, observed in the attached bug:

- A VM is migrating to another host.
- The migration is about to finish.
- VM.destroy call comes and sets POWERING_DOWN status.
- Before the VM is destroyed, the migration finishes successfully.
- The corresponding life cycle event handler sees the VM got down, but
  is not in MIGRATION_SOURCE status and thus the corresponding
  migration cleanup is not initiated.
- The migration source thread is not finished and the outgoing
  migration semaphore is not adjusted.
- When it happens the second time, the outgoing migration semaphore is
  exhausted and no further migrations are possible until Vdsm is
  restarted.

The VM life cycle logic is driven by switching the VM statuses and is
not necessarily bound to the current state of the VM.  For example,
when a migration finishes and the VM gets down on the source host,
Vdsm can see in its status that the VM was migrating although the
migration job is already done.  So using other checks at critical
points would require tracking means similar to the VM status rather
than just examining the current libvirt state of the VM.
Additionally, there is a danger of causing confusion at Engine side.

For these reasons, this patch allows switching the VM status to
POWERING_DOWN only when the VM is UP, when it is completely safe to do
so.

In theory, it may not be without caveats.  For example, if a migration
fails while the VM is powering down, we end up in UP state rather than
POWERING_DOWN.  But this is (also) a corner case and at worst Engine
will allow issuing VM power off again, which shouldn't cause any harm,
unlike not executing actions bound to special VM statuses.  This could
be improved, if needed, in further patches, where Vm._guestEvent could
be considered.

An additional protection could be to cancel a migration when VM
shutdown or destroy is called.  But as for shutdown, Engine doesn't
call shutdown during migration and calls power off instead.  And as
for destroy, it could be dangerous to attempt additional actions
there, especially when calling destroy is part of the migration flow
itself.  Finally, a migration cannot be always canceled, most notably
when the migration is in post copy phase.

Change-Id: Ia2685391d1da3027d7bde710512ab20a70f042e3
Bug-Url: https://bugzilla.redhat.com/1959436
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>