tests: add basic test for pvmove
pvmove was lacking test coverage, adding simple test for
pvmove with data verification before and after the move.

This test works in container env but not in mocki env because dm-mirror
kernel module is missing and causes pvmove to fail. Test is marked to
skip on ovirt ci because of that.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1949059
Change-Id: I9d675e80e2f045d9462554110a9f57ca115425dd