disable lvmpolld for pvmove to fix LUN reduce flow
LUN reduce currently fails when LUN is not empty because pvmove
has to be run in order to move the data to different PV. The pvmove
uses specific filter on command line, rejecting all devices except the
ones suitable for data move.

This filter is not passed to lvmpolld and lvmpolld sees only filter
defined in lvm.conf. So when filter in lvm.conf is set to reject more
devices than the pvmove needs the operation fails because lvmpolld does
not see those devices.

This can be solved using devices file feature of lvm which should be
available later in RHEL8.5 and RHEL9.0. At this point we can only
disable lvmpolld for pvmove operation to provide a quick solution.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1949059
Change-Id: I2d074fa84b4881d6749038c44ce6ea39d3c4d8a2