sp: Watch the SPM lease
After acquiring the SPM lease, start watching the lease status, and
panic if the status is not expected. Before releasing the SPM lease,
stop the watchdog.

The watchdog is used only when sanlock supports inquire(). This version
is available now in Centos stream, and will be available soon in RHEL
8.4.z.

Starting and stopping the watchdog panic on errors:
- If we fail to start the watchdog, we cannot be sure that the lease
  remain value.
- If we fail to stop the watchdog, the watchdog will panic when we
  release the lease.

Failure to start or stop the watchdog is very unlikely. If we cannot
start or stop a thread, we probably have much bigger issue.

The watchdog is enabled by default, but can be disabled if it causes
trouble. The check interval is 20 seconds, matching sanlock check
internal (io_timeout * 2).

To change the configuration use this drop-in file:

$ cat /etc/vdsm/vdsm.conf.d/99-local.conf
[spm]
watchdog_enable = true
watchdog_interval = 20

Change-Id: Iba7bf5b620371e991f62a08b13f378dcb44035e2
Bug-Url: https://bugzilla.redhat.com/1961752
Signed-off-by: Nir Soffer <nsoffer@redhat.com>