spec: Require libvirt 7.0.0-14
This version fixes critical bug[1] in live storage migration, where after a
"successful" operation libvirt forgets the target disk and all
operations on the vm will fail.

Since we have support only RHEL (like) distros now, and the package
should exists in Centos stream, remove the complicated branches.

The package is available in Centos stream advancedvirt-common repo:
http://mirror.centos.org/centos/8-stream/virt/x86_64/advancedvirt-common/Packages/l/

The advancedvirt-common repo was added to check-patch.repos. I'm not
sure this is the right way since the build will fail if the remote repo
is not accessible, but lets see if this actually works.

[1] https://bugzilla.redhat.com/1955667

Change-Id: I45b4d3275e2c10f6b9dd39e8e5fe8b2aee4cbb48
Signed-off-by: Nir Soffer <nsoffer@redhat.com>