tests: Fix tests checking sparse images
Some tests assumed that qemu allocates at least one file system block,
but this assumption is not correct for current qemu-img. On XFS it may
set the file system extent size to 1 MiB, which increase the minimal
allocation size.

Other test wrote small amount of data which makes it hard to predict
the total allocation. Change the test to write 1 MiB of data to make
behave in the same way on different file systems.

Change-Id: Ifdc2af605de35e33bb22bfdc83f867ea97a26a95
Signed-off-by: Nir Soffer <nsoffer@redhat.com>