clear_bitmaps.py: introduce clear_bitmaps job
When a snapshot is restored, all the bitmaps should be removed
from the volume chain.

Removing the bitmaps one by one using the remove_bitmap job can
flood the engine and VDSM with jobs, so a new job is needed for
removing the bitmaps in an efficient way.

The engine will use this new job to remove all the bitmaps
from a volume, and will use it for the entire chain when
a snapshot is restored.

Change-Id: Idbb96121ed23573a902433a00b89771d028ddc63
Bug-Url: https://bugzilla.redhat.com/1952577
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>