automation: Simplify pip requirements
We installed virtualenv using pip, requiring virtualenv<20. This was
probably required when we had to build both python 2 and 3. Currently
installing virtualenv<20 fails with:

    Could not find a version that satisfies the requirement
    virtualenv<20 (from versions: )

The error may be caused by network outage, but it reveals that we are
using unneeded requirements.

Python 3 has venv module, use it instead virtualenv.

We also require coverage>=5. I don't remember why we added this, lets
use the latest version.

Change-Id: I7245e745c481c04e9807bdef5e70ccbed402a413
Signed-off-by: Nir Soffer <nsoffer@redhat.com>