Automatically fix bad ownership/access mode of vdsm log files
Currently the vdsm service will fail to start if it can't write to
the vdsm log file. It can happen i.e. due to bad logrotate, that
keeps the vdsm log file being owned by the root.

This patch adds checks to verify that both file owners and access
mode are set correctly for both vdsm and mom log files.

Change-Id: Ibc2c113d6c973187594672947af231b87914ab0e
Signed-off-by: Lev Veyde <lveyde@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1970008