tests: Minimize use of imageio internals
To test dirty bitmaps, we used ovirt_imageio._internal.nbd.Extent.flags.
This is internal detail which is not part of imageio public API, and it
may change between releases.

imageio 2.2.0-1 changes the NBD_STATE_HOLE bit (1) to EXTENT_DIRTY (2).
Tests using this value breaks with newer imageio version.

Change the test to access the extent properties (length, dirty) to make
the test work with both old and new imageio.

Change-Id: Iadba41b3b6fb6f1f54e643f40739ff4d64ff343f
Bug-Url: https://bugzilla.redhat.com/1971182
Signed-off-by: Nir Soffer <nsoffer@redhat.com>