volumemetadata: improve parsing of storage values
Currently the _lines_to_dict() function can fail when decoding a value
from storage if the value is not valid which causes the remaining
values get skipped.

Add a try/except block where save errors and skip the invalid values.

If there is a skipped value it is later treated as missing key and
saved errors are logged later in volumemetadata.dump()

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1870887
Change-Id: Ib7f7c27cf7c7ff3ebf8b9ce799f7654f6f79ed41