image: Fix measure of block volume in copyCollpase
Since qemu 5.1 we need to use "host_device" driver when accessing block
devices. This was added vdsm 4.40.33 in:

commit 1718b8784e841405574c44abe2357997e3235723
CommitDate: Fri Oct 2 08:21:15 2020 +0000

    storage: support measuring without backing file

But we missed a second call to qemuimg.measure() in copyCollapsed().
When creating a template from vm without a parent volume, the operation
fails with:

    vdsm.common.cmdutils.Error: Command ['/usr/bin/qemu-img', 'measure',
    '--output', 'json', '-O', 'qcow2', 'json:{"file": {"driver": "file",
    "filename": "/path..."}, "driver": "qcow2"}'] failed with rc=1 out=b''
    err=b'qemu-img: Could not open \'json:{"file": {"driver": "file",
    "filename": "/path..."}, "driver": "qcow2"}\': \'file\' driver
    requires \'/path...\' to be a regular file\n'

Copying volume with a parent is not effected, since in this flow we use
another way to estimate the destination volume size.

Fixed by adding "block" boolean key in Volume.getVolumeParams(), and
using it to call qemuimg.measure() correctly when measuring a block
volume.

is_block() was moved up to VolumeManifest() to make pylint happy. The
bad tests faking volume parameters update to report also the new "block"
key.

Change-Id: I8169b7a1c733fd8633cb265436d433ab83961ef5
Bug-Url: https://bugzilla.redhat.com/1973345
Signed-off-by: Nir Soffer <nsoffer@redhat.com>