Revert "spec: Exclude broken libvirt version"
This reverts commit fc75836c5468c81bdf540bcc6101f485df5e235d.

The original patch was needed for these bugs:

- https://bugzilla.redhat.com/1966842
  Fixed upstream

- https://bugzilla.redhat.com/1971182
  Fixed in vdsm master an imageio 2.2.0

- https://bugzilla.redhat.com/1970277
  Relatively small bug that's not worth blocking recent libvirt and
  qemu.

The current limit on libvirt making it harder to test vdsm with current
libvirt and qemu version in RHEL 8.5.

Finally, the original patch does not work in OST.

If we want to limit libvirt version on Centos Stream, it should be done
in the stable branch, not in master.

Change-Id: I56ad2bb2af386ac2534c5787111a56487d0e60d9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>