tests: fileutils: Fix test on python 3.6
The test wrongly checked for EPERM, while actual error is EACCES.  The
test does not run it the CI since it requires unprivileged user, and CI
runs as root. Developers usually run the tests with newer python that
does not raise in this case.  Running the tests on a RHEL 8.5 host as
unprivileged user revealed the issue.

Change-Id: Ida98351e9d8726756ff5cd15e41ea8ab7dbe6e5b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>