drivemonitor: Allow setting threshold with index
When setting threshold for a normal drive, we always use the top layer,
so we register threshold for the drive name (vda). In historic libvirt
versions, there was no other way to refer to the top layer, but now we
have an indexed name also for the top layer.

When registering the top layer using the drive name, recent libvirt
(reported in 6.6) submits block threshold event twice; once for the
registered name (vda) and once for the indexed name (vda[7]).

The unexpected event logs a warning like:

    2021-04-10 08:12:10,550+0800 WARN  (libvirt/events) [virt.vm]
    (vmId='7af5361c-012b-41f9-9ea9-33917df23f21') Unknown drive 'sdi[7]' for
    vm 7af5361c-012b-41f9-9ea9-33917df23f21 - ignored block threshold event
    (drivemonitor:177)

Submitting the event twice in libvirt is likely a bug in libvirt, but
the best way to avoid this is to always use indexed name in vdsm.

Using indexed name is also the only way we can monitor block threshold
on scratch disks and block copy destination image.

This change adds an optional index argument to set_threshold(), similar
to the index argument in clear_threshold(). This allows callers to start
using index also for drive top layer.

Using an index, we will get only one event for the indexed name,
eliminating the warning. However since we get indexed name, we must
parse the name to drive name and index to locate the drive.

A minor issue is that we cannot log the path when since we don't have a
way to get the path related to the libvirt index. We may add this later
to improve logging.

Change-Id: Ia5c3b15410b6649d015216c69490576daa12a43e
Bug-Url: https://bugzilla.redhat.com/1948177
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>