virt: report CPU topology
Report detailed information about CPU topology. XML with libvirt
capabilities contains topology information in following form:

    <topology>
        <cells num='2'>
          <cell id='0'>
            <cpus num='4'>
              <cpu id='0' socket_id='0' die_id='0' core_id='0' siblings='0-1'/>
              ...
            </cpus>
          </cell>
          ...
        </cells>
    </topology

We parse the content so that we can pass it to Engine.

We now query the libvirt capabilities on every Host.getCapabilities
call. Because the results of numa._numa() are cached and the cache is
never cleared this may lead to VDSM consuming and never freeing a lot of
memory in the long run. This issue will be addressed in a separate
patch.

Change-Id: Ifc5a4db88eb998d73fe55df6c92170eface57e69
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>