virt: storage: Support top volume index
Since libvirt 5.1.0[1] we have also an index for the top layer. Update
disk xml parsing to extract the index from the source element.

Since we have to get the index from source element for the top volume,
and the backingStore element for the backing chain, I had to rewrite the
parsing code.

The new parsing code is more strict, enforcing the libvit domain xml
format. For example, we used to allow missing <backingStore/> element,
but libvirt domain xml does not specify this element as optional.

With this change, we always get a target name for the top volume, so in
blockCopy we always use indexed name instead of None.

[1] https://bugzilla.redhat.com/1451398

Change-Id: I58038d94ccc72d9d39c4cd514af6b14b33d1dfba
Bug-Url: https://bugzilla.redhat.com/1948177
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>