vm: Always use index for setting threshold
Always use the indexed name when setting block threshold. This does not
eliminate the double block threshold events[1], but we can drop the
event for the drive name safely.

Update the tests to verify that we always register events for indexed
name.

Update the tests to trigger the double events like libvirt and verify
that we ignore the event with the drive name and handle the event for
the indexed name.

[1] https://bugzilla.redhat.com/1983429

Change-Id: Ia78c2fa4374d643e35a42ca3b78ed6fcc323ac30
Bug-Url: https://bugzilla.redhat.com/1948177
Signed-off-by: Nir Soffer <nsoffer@redhat.com>