virt: improve the check when disk should be refreshed
Improve the check when the disk during migration process should be
refreshed on the destination host. Currently condition is met also
when the migration thread waits for migration semaphore and can result
into unnecessary refresh failure as the VM is not defined on the
destination host by that time. New condition is met only when VM is
defined on the destination host.

For this reason add new method into migration thread which returns True
if the VM is defined on the destination host.

Also add new VM method which determines if the disk refresh is needed or
not. Besides nicer code flow, it reduce the need of adding migration
thread into fake VM classes in the tests.

For recovery flows set _state to STARTED, which further simplifies the
condition and only _state needs to be checked.

Change-Id: I7a8ceccc04f2760c978ad0c9fee62d6f86cede41
Bug-Url: https://bugzilla.redhat.com/1883399
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>