storage: support preparing illegal volumes
As part of the effort to allow disk format conversion, we need to
support creation and preparation of illegal volumes.

When converting the format, an additional volume will be created under
the same image, this disk will serve as the target disk for the
conversion which will be done using SDM.copy_data. The target will be
created illegal, and will remain so until the conversion is complete.

This patch handles the preparation of the an illegal volume.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: I342cccad51749531936702d404b30c75f5dbc54f