storage: support creating illegal volumes
As part of the effort to allow disk format conversion, we need to
support creation and preparation of illegal volumes.

When converting the format, an additional volume will be created under
the same image, this volume will serve as the target volume for the
conversion which will be done using SDM.copy_data. The target will be
created illegal, and will remain so until the conversion is complete.

This can also potentially simplify the disk upload flow by removing an
API call that changes the volume's legality and instead does this during
creation.

This patch adds the ability to control the legality of a volume in
public APIs.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: Ibe6d58c2bd1494602c7faf5dd83762139db64e23