net, tests: Unify nic naming in integration tests
- Instead of having fixed bridge name "br1" in bridge_test
use randomized name with prefix vdsm-. The hebrew bridge in
ipwrapper_test already had randomized suffix, add vdsm- prefix
so it is clearer that the bridge was created by tests.
Update the regex for cleaning up leftover devices to minimize
risk of deleting device that should be present on the node.

- Remove test-network constant that was unused.

Change-Id: Ie2781633aaf7214159da5eaa2b5014827254dde3
Signed-off-by: Ales Musil <amusil@redhat.com>