virt: don't raise when setting disk threshold and VM is not running
If we refresh the disk during migration, setting the threshold may fail
if the destination VM is not running yet. The same situation can happen
when VM is shutting down. Skip setting disk threshold when underlying
libvirt domain is not connected yet and set it to BLOCK_THRESHOLD.UNSET
so it will be eventually set once VM is running by periodic job. This is
the case once migration is finished - disk thresholds are set by
periodic job right after it.

Change-Id: I4012ddcee4b04409d750841536982c84db96e1e2
Bug-Url: https://bugzilla.redhat.com/1981307
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>