virt: don't log error during disk extension when VM is not running
When disk extension is requested, setting new disk threshold can fail
as VM may not be running by that time. E.g. in case if VM migration this
can happen when disk extension is requested before migration is finished
but migration finishes before disk extension - in this case source VM is
already shut down. See stack trace bellow to make it more clear (note:
the stack trace is not present in the log, it was added by modifying
test code, only error message is present in the logs).

Don't log error when libvirt domain is already shut down, as it's
expected disk threshold cannot be set in such case. In this case it
should be safe to ingnore libvirt VIR_ERR_OPERATION_INVALID, see
https://listman.redhat.com/archives/libvirt-users/2021-August/msg00002.html

Example trace:

2021-07-14 15:58:46,438-0400 ERROR (mailbox-hsm/3) [virt.vm] (vmId='2dad9038-3e3a-4b5e-8d20-b0da37d9ef79') Failed to set block threshold for drive 'sdc' (): Requested operation is not valid: domain is not running (drivemonitor:134)
Traceback (most recent call last):
  File "/usr/lib/python3.6/site-packages/vdsm/virt/drivemonitor.py", line 122, in set_threshold
    self._vm._dom.setBlockThreshold(drive.name, threshold)
  File "/usr/lib/python3.6/site-packages/vdsm/virt/virdomain.py", line 101, in f
    ret = attr(*args, **kwargs)
  File "/usr/lib/python3.6/site-packages/vdsm/common/libvirtconnection.py", line 131, in wrapper
    ret = f(*args, **kwargs)
  File "/usr/lib/python3.6/site-packages/vdsm/common/function.py", line 94, in wrapper
    return func(inst, *args, **kwargs)
  File "/usr/lib64/python3.6/site-packages/libvirt.py", line 2609, in setBlockThreshold
    raise libvirtError('virDomainSetBlockThreshold() failed')
libvirt.libvirtError: Requested operation is not valid: domain is not running

Change-Id: I0e81e8a9c47b5c886c73a4fec4a580f05ada0245
Bug-Url: https://bugzilla.redhat.com/1981079
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>