virt: skip VM resume in cases when it's not supported
After disk extend we always call Vm.cont() to eventually resume VM if
it was paused due to ENOSPC before disk extend was finished. However,
VM can be in a state when resume is not supported and it's status
shouldn't be changed (e.g. VM is being migrated or is shutting down).
Check the VM status before calling resume and skip it in these cases.

Change-Id: Ifdd8f6088817a1f07c8e012cd8fae57c05815dd5
Bug-Url: https://bugzilla.redhat.com/1981079
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>