tests: Add missing state for extending drives
We want to always use the index when setting block threshold, but the
current test rig does not build enough fake state to do this.

To use index we need to find the drive actual volume chain, using drive
alias. So we need to set drive.alias when creating a drive.

To lookup drive by alias, we need to have the right disk xml in the fake
domain. This xml must also have the index and path.

To find drive volume id, we need drive volumeChain. We generated a
partial volume chain with enough info to satisfy the code using it.

Drive.index is not related to libvirt index reported in the xml, but for
simplicity we use the same value, since we already abuse the index for
creating unique path.

To manage disk xml, the fake domain keeps now the devices xml element.
This allows easy modification of libvirt xml for adding and removing
drives or simulating changes in libvirt xml.

Change-Id: Iec2011bee16200759dedae9f6a07bb7280c748b9
Bug-Url: https://bugzilla.redhat.com/1948177
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>