vm: Move up handling of NotConnectedError
If a VM is not running when calling DriveMonitor.set_threshold(), we
swallowed the error. This makes it harder to avoid bogus errors and
warning in the rest of the code.

We want to call Vm._drive_get_actual_volume_chain() in
Vm._update_drive_volume_size(). This call may fail in the same way when
the VM is not running when self._dom is virdomain.Defined() or
virdomain.Disconnected(). We can handle NotConnectedError around this
call, but handing this error everywhere is not attractive solution.

Handling expected conditions like non-running VM should be done at the
top level, since we don't have anything to do except logging a debug
message.

Change DriveMonitor.set_threshold() to handle any error and unset the
drive threshold_state, but raise the error so callers can handle it.
This is more correct and removes the dependency on the virdomain module.

Handle NotConnectedError in Vm.refresh_disk() around the call to update
drive volume size, so we don't fail the call during migration.

Handle NotConnectedError in Vm.after_volume_extension(), so we abort the
extend completion flow once we detect that the VM is not running. This
also skip the call to resume the VM, which log other bogus warnings.

Change-Id: I79bdcf16474aa7e12c5e9e32948bc0b562c29446
Bug-Url: https://bugzilla.redhat.com/1948177
Signed-off-by: Nir Soffer <nsoffer@redhat.com>