vm: Remove pointless getVolumeSize() call
When refreshing a disk, we called getVolumeSize() twice, once before
updating drive volume size, and once after the update. The second call
is not needed since volume size cannot change by
Vm._update_drive_volume_size(). Even if we are running on the SPM, and
the volume size was extended again, we care only about the size after
the refresh.

Change-Id: I79d7d8275a5ff938d0c95cdb9490add83fa5e734
Signed-off-by: Nir Soffer <nsoffer@redhat.com>