drivemonitor: Always use block threshold events
When we added support for block threshold events in 2017, we made it
configurable so we can disable the feature if we find an issue in
production. After using this for 4 years, we are pretty confident in
block threshold events, and we are pretty sure that disabling them does
not really work since nobody is testing this.

Removing the configuration simplifies the code and the tests, removing
many tests verifying the old behavior, and simplifying the test helpers.

This change only removes the configuration, additional cleanup will
follow.

Change-Id: Ia19cd2904e68a8236cfea3e55ff954667bc870a2
https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>