tests: Convert drivemonitor test to pytest
Convert test parameters to pytest.param() so we can use an id instead of
comment to describe the test case. This makes the test names nicer:

    $ tox -e virt tests/virt/drivemonitor_test.py -- -v
    ...
    test_monitored_drives[non_chunk_drives] PASSED

It also makes it possible to select tests by test id:

    $ tox -e virt tests/virt/drivemonitor_test.py -- -k _disabled
    ...
    test_monitored_drives_flag_disabled[both_drives_enabled] PASSED
    test_monitored_drives_flag_disabled[first_drive_disabled] PASSED
    test_monitored_drives_flag_disabled[second_drive_disabled] PASSED
    test_monitored_drives_flag_disabled[both_drives_disabled] PASSED

Note that I trimmed the real output from pytest, since the long test
names do not play well with gerrit UI. The next patch will improve this
issue.

Reformat the drive configuration properly to make it much more readable.

We used VdsmTestCase only for its logger, but but we can the fake vm
logger.

Change-Id: Ia400e615dc15ccb727351fa2d3e8490a45ab94b3
Signed-off-by: Nir Soffer <nsoffer@redhat.com>